requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
jinja2==3.1.6
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.2.1
python-multipart==0.0.20